+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+* BREAKING CHANGE: Input is sanitized by default, to use unsafe HTML initialize with a sanitize option of false
+* Allow sanitize option on remove invalid HTML from source input
+* BREAKING CHANGE: Remove `to_sanitized_html` method in favour of `sanitize` option on initialize
+* BREAKING CHANGE: Remove `to_sanitized_html_without_images` as no apps use this anymore
+
 ## 5.9.1
 
 * Don't render `[Image: {file-name}]` within a paragraph to avoid invalid HTML